ref: reorganize the location of the hotfix function
